Extract all post publish functionality from the publisher class.
